+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/artifactory/ArtifactoryUploader.java
@@ -5,7 +5,9 @@ import com.teamscale.client.StringUtils;
 import com.teamscale.jacoco.agent.options.ArtifactoryConfig;
 import com.teamscale.jacoco.agent.upload.HttpZipUploaderBase;
 import com.teamscale.report.jacoco.CoverageFile;
+import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
+import okhttp3.Request;
 import okhttp3.ResponseBody;
 import retrofit2.Response;
 
@@ -18,6 +20,11 @@ import java.util.List;
  * Uploads XMLs to Artifactory.
  */
 public class ArtifactoryUploader extends HttpZipUploaderBase<IArtifactoryUploadApi> {
+	/**
+	 * Header that can be used as alternative to basic authentication to authenticate requests against artifactory. For
+	 * details check https://www.jfrog.com/confluence/display/JFROG/Artifactory+REST+API
+	 */
+	public static final String ARTIFACTORY_API_HEADER = "X-JFrog-Art-Api";
 	private final ArtifactoryConfig artifactoryConfig;
 	private String uploadPath;
 
@@ -30,7 +37,12 @@ public class ArtifactoryUploader extends HttpZipUploaderBase<IArtifactoryUploadA
 	@Override
 	protected void configureOkHttp(OkHttpClient.Builder builder) {
 		super.configureOkHttp(builder);
-		builder.addInterceptor(HttpUtils.getBasicAuthInterceptor(artifactoryConfig.user, artifactoryConfig.password));
+		if (artifactoryConfig.apiKey != null) {
+			builder.addInterceptor(getArtifactoryApiHeaderInterceptor());
+		} else {
+			builder.addInterceptor(
+					HttpUtils.getBasicAuthInterceptor(artifactoryConfig.user, artifactoryConfig.password));
+		}
 	}
 
 	@Override
@@ -60,5 +72,13 @@ public class ArtifactoryUploader extends HttpZipUploaderBase<IArtifactoryUploadA
 	@Override
 	public String describe() {
 		return "Uploading to " + uploadUrl;
+	}
+
+	private Interceptor getArtifactoryApiHeaderInterceptor() {
+		return chain -> {
+			Request newRequest = chain.request().newBuilder().header(ARTIFACTORY_API_HEADER, artifactoryConfig.apiKey)
+					.build();
+			return chain.proceed(newRequest);
+		};
 	}
 }